// src/app/evaluation-trim/models/evaluation-setting.ts
import { Year } from '../../establishment/models/year';
import { Evaluation } from './evaluation';

export class EvaluationSetting {
    year: Year;
    eval11: Evaluation;
    eval12: Evaluation;
    eval13: Evaluation;
    eval14: Evaluation;

    eval21: Evaluation;
    eval22: Evaluation;
    eval23: Evaluation;
    eval24: Evaluation;

    eval31: Evaluation;
    eval32: Evaluation;
    eval33: Evaluation;
    eval34: Evaluation;
}